chore: flatten config directory (#94)
