+++ src/abis/l2/index.js
@@ -1,4 +1,4 @@
-import L2_ERC20_ABI from './ERC20.json';
+import L2_ERC20_ABI from './erc20.json';
 import L2_BRIDGE_ABI from './token_bridge.json';
 
 export {L2_ERC20_ABI, L2_BRIDGE_ABI};
