--- src/config/addresses/index.js
@@ -1,3 +1,0 @@
-export * from './contracts/contracts.l1';
-export * from './tokens/tokens.l1';
-export * from './tokens/tokens.l2';
