+++ src/utils/blockchain/starknet.js
@@ -81,3 +81,9 @@ export const getTransactionHash = (
     ...additionalData
   ]);
 };
+
+export const hashEquals = (...data) => {
+  return !!data.reduce((d1, d2) =>
+    starknet.hash.computeHashOnElements(d1) === starknet.hash.computeHashOnElements(d2) ? d1 : ''
+  );
+};
